* gcc.dg/tree-ssa/loop-5.c: Use long instead of int.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@91740 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.dg/tree-ssa/loop-5.c
@@ -3,11 +3,11 @@
 /* { dg-do compile } */
 /* { dg-options "-O1 -fdump-tree-vars" } */
 
-void foo(int);
+void foo(long);
 
 void xxx(void)
 {
-  int iter, jter;
+  long iter, jter;
 
   for (iter = 0, jter = 2; iter < 100; iter++, jter++)
     {
